refactor: ♻️ move and rename functions
