--- src/feasibility_data/common/redcap/__init__.py
@@ -1,1 +1,0 @@
-"""REDCap common functions."""
